Set up masonry layouts via Salvattore.js
This adds a left-to-right masonry grid to the site using Salvattore.js.
This offers one significant advantage over the CSS3 columns approach
because content is ordered left-to-right horizonally instead of
descending down each row; this is generally what people expect.

In some cases (TF page for now) the other approach is still used (order
doesn't matter here), but in both cases the grid spacing is the same.
Salvattore needs negative margins in order to produce even columns, btw.