Implement data-driven press section
Since these entries are more about image content (really a series of
dynamic galleries that expand when clicked), using a blog setup is no
longer appropriate. This commit implements a data-driven approach
instead; press entries are specified as items in a YAML array that lives
in data/press.yml. Each item has an array of one or more images, a
title, and possibly a date (currently not using it but just in case).